requirements.txt
tensorflow-gpu
scikit-image
tqdm
